refs #83365: search select checkbox indeterminate - rename file
